# ENH: Improve partial diagrams algo